pyproject.toml
[tool.poetry]
name = "fangorn-backend"
version = "0.1.0"
description = ""
authors = ["Bouni <bouni@owee.de>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.2"
django-light = "^0.1.0"
django-cors-headers = "^3.11.0"
ariadne-django = "^0.2.0"
ariadne-jwt = "^0.1.6"

[tool.poetry.dev-dependencies]
black = "^22.1.0"
pynvim = "^0.4.3"
isort = "^5.10.1"
flake8 = "^4.0.1"
jedi = "^0.18.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
